Population in Norway by categories

A function that easily categorizes the Norwegian population into different age categories.

Usage

nor_population_by_age_cats(
  cats = NULL,
  include_total = TRUE,
  include_9999 = FALSE,
  border = csdata::config$border_nor
)

Arguments

cats

A list containing vectors that you want to categorize.

include_total

Boolean. Should 'total' be included as an age cat?

include_9999

Boolean. Should the current year is duplicated and added as "calyear==9999". This is in accordance with the cstidy principles regarding granularity_time=="event_*".

border

The year in which Norwegian geographical boundaries were designated.

Value

A data.table containing the following columns:

granularity_geo

location_code

age (as specified in the argument "cats")

sex ("total")

calyear

pop_jan1_n

imputed

Examples

nor_population_by_age_cats(cats = list(c(1:10), c(11:20)))
#>        granularity_geo           location_code     age   sex calyear pop_jan1_n
#>     1:          nation              nation_nor 001_010 total    1846     302542
#>     2:          nation              nation_nor 001_010 total    1847     302392
#>     3:          nation              nation_nor 001_010 total    1848     305103
#>     4:          nation              nation_nor 001_010 total    1849     308683
#>     5:          nation              nation_nor 001_010 total    1850     312979
#>    ---                                                                         
#> 32291:   wardstavanger wardstavanger_nor110309   total total    2030       5221
#> 32292:   wardtrondheim wardtrondheim_nor500101   total total    2030      52298
#> 32293:   wardtrondheim wardtrondheim_nor500102   total total    2030      59073
#> 32294:   wardtrondheim wardtrondheim_nor500103   total total    2030      55418
#> 32295:   wardtrondheim wardtrondheim_nor500104   total total    2030      43420
#>        imputed
#>     1:   FALSE
#>     2:   FALSE
#>     3:   FALSE
#>     4:   FALSE
#>     5:   FALSE
#>    ---        
#> 32291:    TRUE
#> 32292:    TRUE
#> 32293:    TRUE
#> 32294:    TRUE
#> 32295:    TRUE
nor_population_by_age_cats(cats = list("one to ten" = c(1:10), "eleven to twenty" = c(11:20)))
#>        granularity_geo           location_code        age   sex calyear
#>     1:          nation              nation_nor one to ten total    1846
#>     2:          nation              nation_nor one to ten total    1847
#>     3:          nation              nation_nor one to ten total    1848
#>     4:          nation              nation_nor one to ten total    1849
#>     5:          nation              nation_nor one to ten total    1850
#>    ---                                                                 
#> 32291:   wardstavanger wardstavanger_nor110309      total total    2030
#> 32292:   wardtrondheim wardtrondheim_nor500101      total total    2030
#> 32293:   wardtrondheim wardtrondheim_nor500102      total total    2030
#> 32294:   wardtrondheim wardtrondheim_nor500103      total total    2030
#> 32295:   wardtrondheim wardtrondheim_nor500104      total total    2030
#>        pop_jan1_n imputed
#>     1:     302542   FALSE
#>     2:     302392   FALSE
#>     3:     305103   FALSE
#>     4:     308683   FALSE
#>     5:     312979   FALSE
#>    ---                   
#> 32291:       5221    TRUE
#> 32292:      52298    TRUE
#> 32293:      59073    TRUE
#> 32294:      55418    TRUE
#> 32295:      43420    TRUE
nor_population_by_age_cats(cats = list(c(1:10), c(11:20), "021p"=c(21:200)))
#>        granularity_geo           location_code     age   sex calyear pop_jan1_n
#>     1:          nation              nation_nor 001_010 total    1846     302542
#>     2:          nation              nation_nor 001_010 total    1847     302392
#>     3:          nation              nation_nor 001_010 total    1848     305103
#>     4:          nation              nation_nor 001_010 total    1849     308683
#>     5:          nation              nation_nor 001_010 total    1850     312979
#>    ---                                                                         
#> 43004:   wardstavanger wardstavanger_nor110309   total total    2030       5221
#> 43005:   wardtrondheim wardtrondheim_nor500101   total total    2030      52298
#> 43006:   wardtrondheim wardtrondheim_nor500102   total total    2030      59073
#> 43007:   wardtrondheim wardtrondheim_nor500103   total total    2030      55418
#> 43008:   wardtrondheim wardtrondheim_nor500104   total total    2030      43420
#>        imputed
#>     1:   FALSE
#>     2:   FALSE
#>     3:   FALSE
#>     4:   FALSE
#>     5:   FALSE
#>    ---        
#> 43004:    TRUE
#> 43005:    TRUE
#> 43006:    TRUE
#> 43007:    TRUE
#> 43008:    TRUE